Move CloudFromation stack viewer to CloudFormation package
