+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/cloudformation/stack/TreeView.kt
@@ -1,6 +1,6 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-package software.aws.toolkits.jetbrains.core.stack
+package software.aws.toolkits.jetbrains.services.cloudformation.stack
 
 import com.intellij.ide.util.treeView.NodeDescriptor
 import com.intellij.openapi.Disposable
