+++ jetbrains-core/tst/software/aws/toolkits/jetbrains/services/cloudformation/stack/Mocks.kt
@@ -1,6 +1,6 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-package software.aws.toolkits.jetbrains.core.stack
+package software.aws.toolkits.jetbrains.services.cloudformation.stack
 
 import org.junit.Assert
 import org.mockito.Mockito
